refactor(client): delete legacy driver stack; slim maestro-client provisioning
Delete the concrete driver stack that implemented the removed Driver interface:
AndroidDriver, IOSDriver, WebDriver, CdpWebDriver, the chromedevtools webview
clients, utils/ScreenshotUtils, android/AndroidLaunchArguments, and
device/DeviceArtifactFiles (all pure-legacy — only the deleted drivers used them).

DeviceService.startDevice is LIVE (maestro start-device + the device-core
selectDevice boot path) and depended on the concrete AndroidDriver (emulator
locale setup) and CdpWebDriver (chromium launch). Those can't survive without
the deleted Driver interface, so the provisioning surface DeviceService needs is
carved into two focused, facade-free helpers: AndroidDeviceLocaleSetter
(install/setLocale/uninstall, behavior-preserving) and WebDeviceLauncher
(detached chromium launch). AndroidDriverSetLocaleTest is repointed to the new
setter as AndroidDeviceLocaleSetterTest.

KEPT as shared plumbing (live device-core / selectDevice depend on them, noted in
the W2 report): DebugLogStore (CLI logging; only used maestro.Driver as a
classloader anchor, swapped to its own class), CapturedDeviceArtifact, and the
android/ adb connection plumbing (AndroidDeviceConnection et al.).

Driver-module test casualties: CdpWebDriverTest, IOSDriverTest, WebDriverTest,
FlutterWebSemanticsIdentifierTest, chromedevtools tests, AndroidLaunchArgumentsTest.
AdbSocketTimeoutTest keeps its adb-transport tests; only its one chromedevtools
case is trimmed.
