--- maestro-client/src/main/java/maestro/android/chromedevtools/ChromeDevToolsClient.kt
@@ -1,9 +1,0 @@
-package maestro.android.chromedevtools
-
-import maestro.TreeNode
-import java.io.Closeable
-
-// Reads on-screen WebView DOMs via Chrome DevTools. Seam so the fetch can be faked in tests.
-interface ChromeDevToolsClient : Closeable {
-    fun getWebViewTreeNodes(): List<TreeNode>
-}
