--- maestro-client/src/main/java/maestro/device/DeviceArtifactFiles.kt
@@ -1,10 +1,0 @@
-package maestro.device
-
-/** Canonical filenames the driver layer writes device artifacts to. */
-object DeviceArtifactFiles {
-    const val SIMULATOR_LOG = "device-simulator.log"
-    const val XCTEST_LOG = "device-xctest.log"
-    const val LOGCAT = "device-logcat.txt"
-    const val CRASH_REPORT = "crash-report.txt"
-    const val ANR_REPORT = "anr-report.txt"
-}
